[EI-254] [web] feat: Add useCreatingCustomForecastIndicator hook for custom forecast indicator creation
